refactor: split registry and rewrites from compiler classes
Smaller modules will help with organization.
