+++ ibis_bigquery/udf/__init__.py
@@ -17,7 +17,8 @@ except ImportError:
 
 from ibis.expr.signature import Argument as Arg
 
-from ..compiler import BigQueryUDFNode, compiles
+from ..compiler import compiles
+from ..operations import BigQueryUDFNode
 from ..datatypes import UDFContext, ibis_type_to_bigquery_type
 from ..udf.core import PythonToJavaScriptTranslator
